+++ mobile/android/android-components/components/browser/toolbar/src/main/java/mozilla/components/browser/toolbar/display/DisplayToolbar.kt
@@ -6,6 +6,7 @@ package mozilla.components.browser.toolbar.display
 
 import android.annotation.SuppressLint
 import android.content.Context
+import android.graphics.Color
 import android.graphics.drawable.Drawable
 import android.view.Gravity
 import android.view.View
@@ -29,10 +30,10 @@ import mozilla.components.browser.toolbar.internal.wrapAction
 import mozilla.components.concept.toolbar.Toolbar
 import mozilla.components.concept.toolbar.Toolbar.SiteSecurity
 import mozilla.components.concept.toolbar.Toolbar.SiteTrackingProtection
+import mozilla.components.concept.toolbar.Toolbar.SiteTrackingProtection.OFF_FOR_A_SITE
+import mozilla.components.concept.toolbar.Toolbar.SiteTrackingProtection.OFF_GLOBALLY
 import mozilla.components.concept.toolbar.Toolbar.SiteTrackingProtection.ON_NO_TRACKERS_BLOCKED
 import mozilla.components.concept.toolbar.Toolbar.SiteTrackingProtection.ON_TRACKERS_BLOCKED
-import mozilla.components.concept.toolbar.Toolbar.SiteTrackingProtection.OFF_GLOBALLY
-import mozilla.components.concept.toolbar.Toolbar.SiteTrackingProtection.OFF_FOR_A_SITE
 
 /**
  * Sub-component of the browser toolbar responsible for displaying the URL and related controls.
@@ -127,7 +128,13 @@ internal class DisplayToolbar(
 
     private var siteTrackingProtection = OFF_GLOBALLY
 
-    internal var securityIcons = SiteSecurityIcons.getDefaultSecurityIcons(context, defaultColor)
+    internal var securityIcon = context.getDrawable(R.drawable.mozac_ic_site_security)
+        set(value) {
+            field = value
+            siteSecurityIconView.setImageDrawable(value)
+        }
+
+    internal var securityIconColor = defaultColor to defaultColor
         set(value) {
             field = value
             setSiteSecurity(currentSiteSecurity)
@@ -160,10 +167,8 @@ internal class DisplayToolbar(
             setOnClickListener(null)
         }
 
-    internal val siteSecurityIconView = AppCompatImageView(context).apply {
+    internal val siteSecurityIconView = SiteSecurityIconView(context).apply {
         setPadding(resources.getDimensionPixelSize(R.dimen.mozac_browser_toolbar_icon_padding))
-
-        setImageDrawable(securityIcons.insecure)
 
         // Avoiding text behind the icon being selectable. If the listener is not set
         // with a value or null text behind the icon can be selectable.
@@ -305,12 +310,17 @@ internal class DisplayToolbar(
      * Sets the site's security icon as secure if true, else the regular globe.
      */
     fun setSiteSecurity(secure: SiteSecurity) {
-        val drawable = when (secure) {
-            SiteSecurity.INSECURE -> securityIcons.insecure
-            SiteSecurity.SECURE -> securityIcons.secure
+        @ColorInt val color = when (secure) {
+            SiteSecurity.INSECURE -> securityIconColor.first
+            SiteSecurity.SECURE -> securityIconColor.second
         }
-        siteSecurityIconView.setImageDrawable(drawable)
+        if (color == Color.TRANSPARENT) {
+            siteSecurityIconView.clearColorFilter()
+        } else {
+            siteSecurityIconView.setColorFilter(color)
+        }
 
+        siteSecurityIconView.siteSecurity = secure
         currentSiteSecurity = secure
     }